requirement.txt
Django==1.5
django-bootstrap3==2.5.6
django-bootstrap3-datetimepicker==2.2.0
django-mailer==0.1.0
django-pipeline==1.3.19
wsgiref==0.1.2
